pyproject.toml
[tool.poetry]
name = "syncshell"
version = "1.0.4"
description = "Keep your machine's shell history synchronized."

license = "MIT"

authors = [
    "Masoud Ghorbani <msud.ghorbani@gmail.com>"
]

readme = "README.md"

repository = "https://github.com/msudgh/syncshell"
homepage = "https://github.com/msudgh/syncshell"

keywords = ["sync", "shell", "history", "bash", "zsh"]

[tool.poetry.dependencies]
python = "^3.7.13"
black = "22.3.0"
fire = "0.1.3"
flake8 = "4.0.1"
halo = "0.0.12"
PyGithub = "1.55"
pytest = "7.1.2"
pytest-testdox = "^3.0.1"

[tool.poetry.dev-dependencies]
pre-commit = "^2.19.0"

[tool.black]
py36 = true
include = '\.pyi?$'
exclude = '''
/(
    \.git
    | dist
)/
'''
